weather/gweather-api-app/requirements.txt
falcon==1.4.1
grpcio==1.12.1
grpcio-tools==1.12.1
gunicorn==19.8.1
protobuf==3.5.2.post1
python-mimeparse==1.6.0
six==1.11.0
